refactor(DEV-9680): extract helpers and types to the separate files
